The MyAwesomePackage Module

Module Index

riskMeasures.CTE2

riskMeasures.EVaR

riskMeasures.EVaR2

riskMeasures.Expectation

riskMeasures.GenCoherent

riskMeasures.GenConvex

riskMeasures.VaR

riskMeasures.distortion

riskMeasures.entropic

riskMeasures.mSD

riskMeasures.meanDeviation

riskMeasures.meanSemiDevi

riskMeasures.meanSemiVariance

riskMeasures.meanVariance

riskMeasures.spectral

Detailed API

riskMeasures.CTE2 — Method

An implementation of the Conditional Tail-Value-at-Risk at level α for a given discrete distribution specified by

states prob

riskMeasures.EVaR — Method

Solving the optimization problem associated with evaluating EVaR as well as providing the dual variables

min (β+log Ee^(t*Y)	)/t  for t>0

riskMeasures.EVaR2 — Method

Solving the optimization problem associated with evaluating EVaR as well as providing the dual variables

min (β+log Ee^(t*Y)	)/t  for t>0

Here with goldenSearch Algorithm

riskMeasures.Expectation — Method

Expectation(states,prob)

Compute the expectation for the random variable given by states under the probability measure given by prob.

riskMeasures.GenCoherent — Method

Coherent risk measure Calculates a general coherent risk measure based on duality representations

for coherent risk measures the conditions E(Z) = 1 and Z ≥ 0 are enforced.
conds needs to be a real valued function F such that F(Z) ≤ 0.
To obtain a coherent risk measure F should be positively homegeneous.

riskMeasures.GenConvex — Method

Convex risk measure Calculates a general convex risk measure based on duality representations

conds must specify the feasible set of the convex conjugate
conjugate must be given via a "simple" expression

riskMeasures.VaR — Method

An implementation of the Value-at-Risk at level α for a given discrete distribution specified by

states prob

riskMeasures.distortion — Method

Distortion risk measure Calculates the distortion risk measure by numerically integrating the distorted distribution function for a given distortion function.

The distribution function is the one for the discrete distribution
implied by states and prob vectors.

riskMeasures.entropic — Method

Entropic risk measure as providing the dual variables

t* log Ee^(Y/t)  for t>0

riskMeasures.mSD — Method

An implementation of the Mean semi-deviation risk measure of order p. This is a coherent risk measure which can be nested by adjusting the risk level β → β⋅√(Δt)

mSD_β^p(Y) = EY + β ⋅ ||(Y-EY)_+||_p

riskMeasures.meanDeviation — Method

Mean Deviation risk Measure of order p

riskMeasures.meanSemiDevi — Method

Mean-upper-semideviation of order p from a target

riskMeasures.meanSemiVariance — Method

Mean-upper-semivariance from a target

riskMeasures.meanVariance — Method

Mean Variance risk Measure

riskMeasures.spectral — Method

Spectral risk measure Calculates spectral risk measure by numerically integrating product of quantile function and given spectral function.

The quantile function is the quantile function for the discrete distribution
implied by states and prob vectors.