RiskMeasures

Julia library for computing risk measures for random variables. The random variable represents profits or rewards that are to be maximized. The computed risk value is also better when greater.

The following risk measures are currently supported

VaR: Value at risk

CVaR: Conditional value at risk

ERM: Entropic risk measure

EVaR: Entropic value at risk

When supported, the risk measure returns also the optimal distribution

General assumptions

Random variables represent rewards (greater value is preferred)

Risk measures become less risk-averse with the increasing value of the risk parameter α or β

Supported distributions

General discrete distributions (DiscreteNonParametric)

Warning: This is package is in development and the computed values should be treated with caution.

Examples

using RiskMeasures
using Distributions

x̃ = DiscreteNonParametric([1, 5, 6, 7, 20], [0.1, 0.1, 0.2, 0.5, 0.1])
 
VaR(x̃, 0.1)   # value at risk
CVaR(x̃, 0.1)  # conditional value at risk
EVaR(x̃, 0.1)  # entropic value at risk
ERM(x̃, 0.1)   # entropic risk measure

We can also compute risk measures of transformed random variables

VaR(5*x̃ + 10, 0.1)   # value at risk
CVaR(x̃ - 10, 0.1)  # conditional value at risk

Extended methods VaR_e, CVaR_e, and EVaR_e also return additional statistics and values, such as the distribution that attains the risk value and the optimal β in EVaR.

See Also

MarketRisk.jl

Functions

Value at Risk

RiskMeasures.VaR — Function

VaR(x̃, α)

Compute the value at risk at risk level α for the random variable x̃.

Risk must satisfy $α ∈ [0,1]$ and α=0.5 computes the median and α=1 computes the essential infimum (smallest value with positive probability) and α=0 returns infinity.

Solves for $\inf \{x ∈ \mathbb{R} : \mathbb{P}[x̃ ≤ x] > 1-α \}$

In general, this function is neither convex nor concave.

RiskMeasures.VaR_e — Function

VaR_e(values, pmf, α; ...)

Compute VaR for a discrete random variable with values and the probability mass function pmf. See VaR(x̃, α) for more details. Also compute the index that achieves

Runs in $n \log(n)$ time where n = length(x̃).

Also returns the value VaR and an index i such that values[i] = x in the minimization above. If such an index does not exist, then returns -1.

Conditional Value at Risk

RiskMeasures.CVaR — Function

CVaR(x̃, α)

Compute the conditional value at risk at level α for the random variable x̃.

The risk level α must satisfy the $α ∈ [0,1]$. Risk aversion increases with an increasing α and, α = 0 represents the expectation, α = 1 computes the essential infimum (smallest value with positive probability).

Assumes a reward maximization setting and solves the dual form

\[\min_{q ∈ \mathcal{Q}} q^T x̃\]

where

\[\mathcal{Q} = \left\{q ∈ Δ^n : q_i ≤ \frac{p_i}{1-α}\right\}\]

and $Δ^n$ is the probability simplex, and $p$ is the distribution of $x̃$.

More details: https://en.wikipedia.org/wiki/Expected_shortfall

RiskMeasures.CVaR_e — Function

CVaR_e(x̃, α)

Compute the conditional value at risk at level α for the random variable x̃. Also compute the equivalent random variable with the same support but a different distribution.

Returns a named tuple with value and the solution random variable that solves the robust CVaR formulation. That is if ỹ is the solution, then the support of x̃ and ỹ are the same and $\mathbb{E}[ỹ] = \operatorname{CVaR}_{α}[x̃]$.

Entropic Value at Risk

RiskMeasures.EVaR — Function

EVaR(x̃, α; βmin = 1e-5, βmax = 100, reciprocal = false, distribution = false)

Compute the EVaR risk measure of the random variable x̃ with risk level α in [0,1].

When α = 0, the function computes the expected value, and when α = 1, then the function computes the essential infimum (a minimum value with positive probability).

The function solves

\[\max_{β ∈ [βmin, βmax]} \operatorname{ERM}_β (x̃) - β^{-1} \log (1/(1-α)).\]

Large values of βmax may cause the computation to overflow.

If reciprocal = false, then the quasi-concave problem above is solved directly. If reciprocal = true, then the optimization is reformulated in terms of λ = 1/β to get a concave function that can be solved (probably) more efficiently

The function implicitly assumes that all elements of the probability space have non-zero probability.

Returns the EVaR and the value β that attains the maximum above.

See: Ahmadi-Javid, A. “Entropic Value-at-Risk: A New Coherent Risk Measure.” Journal of Optimization Theory and Applications 155(3), 2012.

RiskMeasures.EVaR_e — Function

EVaR_e(values, pmf, α; ...)

Compute EVaR for a discrete random variable with values and the probability mass function pmf. See EVaR(x̃, α) for more details.

Entropic Risk Measure

RiskMeasures.ERM — Function

ERM(x̃, β [, x̃min, check_inputs = true])

Compute the entropic risk measure of the random variable x̃ with risk level β.

The optional x̃min parameter is used as an offset in order to avoid overflows when computing the exponential function. If not provided, the minimum value of x̃ is used instead.

Assumes a maximization problem. Using β=0 computes expectation and β=Inf computes the essential infimum (smallest value with positive probability).

More details: https://en.wikipedia.org/wiki/Entropicriskmeasure

RiskMeasures.softmin — Function

softmin(x̃, β; x̃min check_inputs = true)

Compute a weighted softmin function for random variable x̃ with risk level β. This can be seen as an approximation of the arg min function and not the min function.

The operator computes a distribution p such that

\[p_i = \frac{e^{-β x_i}}{\mathbb{E}[e^{-β x̃}]}\]

The optional x̃min parameter is used as an offset in order to avoid overflows when computing the exponential function. If not provided, the minimum value of x̃ is used instead.

The value β must be positive

Essential Infimum

RiskMeasures.essinf — Function

essinf(x̃; distribution = false)

Compute the essential infimum of the random variable x̃, which is the minimum value with positive probability

RiskMeasures.essinf_e — Function

essinf_e(values, pmf; ...)

Compute the value for a random variable with values and the probability mass function pmf.

See essinf_e(x̃) for more details.