RiskMeasures

Documentation for RiskMeasures.

RiskMeasures.AVaR

RiskMeasures.CTE

RiskMeasures.EVaR

RiskMeasures.EVaR2

RiskMeasures.Expectation

RiskMeasures.GenCoherent

RiskMeasures.GenConvex

RiskMeasures.VaR

RiskMeasures.distortion

RiskMeasures.entropic

RiskMeasures.mSD

RiskMeasures.meanSemiDevi

RiskMeasures.meanSemiVariance

RiskMeasures.meanVariance

RiskMeasures.spectral

RiskMeasures.AVaR — Method

AVaR(states,prob,alpha::Float64)

Solves the optimization problem associated with the primal formulation of the Average Value-at-Risk:

\[AVaR_\alpha(Y) = \min_{x\in \mathbb{R}} x + \frac{1}{1-\alpha} \mathbb{E} \left( Y - x \right)_+,\]

where $Y$ is the discrete random variable defined by states and prob.

RiskMeasures.CTE — Method

CTE(states,prob,alpha::Float64)

implements the Conditional Value-at-Risk at level $\alpha$ defined by

\[CTE_\alpha (Y) = VaR_\alpha(Y) + \frac{1}{1-\alpha} \mathbb{E} \left( Y- VaR_\alpha (Y) \right)_+ ,\]

for the random variable $Y$ defined by states and prob.

RiskMeasures.EVaR — Method

EVaR(states,prob,beta::Float64)

Solves the optimization problem associated with the primal formulation of the Entropic Value-at-Risk:

\[EVaR_\alpha(Y) = \min_{x >0} \frac{1}{x} \left( \beta + \log\mathbb{E} e^{xY} \right),\]

where $Y$ is the discrete random variable defined by states and prob. Here, the optimization is done using JuMP and Ipopt.

RiskMeasures.EVaR2 — Method

EVaR2(states,prob,beta)

Solves the optimization problem associated with the primal formulation of the Entropic Value-at-Risk:

\[EVaR_\alpha(Y) = \min_{x >0} \frac{1}{x} \left( \beta + \log\mathbb{E} e^{xY} \right),\]

where $Y$ is the discrete random variable defined by states and prob. Here the optimization is done via the goldenSearch optimization routine implemented as part of this package.

RiskMeasures.Expectation — Method

Expectation(states,prob)

Computes the expectation for the random variable $Y$ with values given by states under the probability measure $Q$ given by prob:

\[\mathbb{E} Y = states \cdot prob\]

RiskMeasures.GenCoherent — Method

GenCoherent(states, prob,conds)

implements a generic coherent risk measure based on the dual representation

\[\rho(Y) := \sup \left( \mathbb{E} YZ \colon \mathbb{E}Z=1, Z\geq 0, F(Z) \leq c \right),\]

where $F(\cdot)$ is a positively homegeneous and convex function and c is a constant.

RiskMeasures.GenConvex — Method

GenConvex(states, prob, conds, conjugate)

implements a generic coherent risk measure based on the dual representation

\[\rho(Y) := \sup \left( \mathbb{E} YZ - conjugate(Z) \colon \mathbb{E}Z=1, Z\geq 0 \right),\]

where $conjugate(\cdot)$ is a and convex function and conds describes the domain of $conjugate(\cdot)$.

RiskMeasures.VaR — Method

VaR(states,prob,alpha::Float64)

implements the Value-at-Risk at level $\alpha$ defined by

\[VaR_\alpha (Y) = \arg \min_x \left( x\in \mathbb{R} : F_Y(x) \geq \alpha \right),\]

for the random variable $Y$ defined by states and prob.

RiskMeasures.distortion — Method

distortion(states, prob,dist)

implements the distortion risk measure

\[\rho(Y) := \int_0^\infty 1- dist(F_Y(x)) dx - \int_{-\infty}^0 dist(F_Y(x) dx,\]

where $Y$ is the discrete random variable defined by states and prob and $dist(\cdot)$ denotes the distortion function which is a right-continuous distribution function.

RiskMeasures.entropic — Method

entropic(states, prob,theta::Float64)

implements the entropic risk measure defined as

\[\rho_\theta(Y) = \theta \cdot \log \mathbb{E} e^{\frac{Y}{\theta}},\]

where $\theta$ is greater 0 and $Y$ is the random variable defined by states and prob.

RiskMeasures.mSD — Method

mSD(states,prob,beta::Float64,p::Float64)

implements the mean semi-deviation of order p which is a coherent risk measure defined by

\[mSD_\beta^p (Y) = \mathbb{E} Y + \beta \lvert \left( Y - \mathbb{E}Y \right)_+ \rvert_p,\]

for the random variable $Y$ defined by states and prob.

RiskMeasures.meanSemiDevi — Method

meanSemiDevi(states, prob,c::Float64,target::Float64,p::Float64)

implements the mean upper-semi variance risk measure of order $p \geq 1$ from a target $t$ defined by

\[\rho_{c,t}^p(Y) = \mathbb{E}Y + c \cdot \lVert \left( Y - t \right)_+ \rVert_p ,\]

where $c >0$ and $Y$ is the random variable defined by states and prob.

RiskMeasures.meanSemiVariance — Method

meanSemiVariance(states, prob,c::Float64,t::Float64)

implements the mean upper-semi variance risk measure from a target $t$ defined by

\[\rho_{c,t}(Y) = \mathbb{E}Y + c \cdot \mathbb{E} \left( Y - t \right)^2_+ ,\]

where $c >0$ and $Y$ is the random variable defined by states and prob.

RiskMeasures.meanVariance — Method

meanVariance(states, prob,c::Float64)

implements the mean Variance risk measure defined by

\[\rho_c(Y) := \mathbb{E}Y + c \cdot \mathbb{E} \left( Y- \mathbb{E}Y)^2 \right),\]

where $c >0$ and $Y$ is the random variable defined by states and prob.

RiskMeasures.spectral — Method

spectral(states, prob, spec)

implements the spectral risk measure

\[\rho(Y) := \int_0^1 spec(x) \cdot F_Y^{-1}(\alpha) d\alpha,\]

where $Y$ is the discrete random variable defined by states and prob and $spec(\cdot)$ denotes the spectral function which is

increasing

integrates to 1.